Refactor-4 Move llm specific functions to another file.
